+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## Unreleased
+
+### Added
+
+- licenses can be removed from a device on the moreInfo screen instead of having to go back into Licenses
+
 ## 4.1
 
 ### Fixed